Move ByNameClosure earlier in the pipleline
 - Make it use changeOwnerAfter
 - Factor out TransformByNameApply as a common base trait
   between it and ElimByName
